# wardrobe/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wardrobe API",
        "description": "API Documentation for Wardrobe BE",
        "contact": {
            "email": "flazen.edu@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/dct/{id}": {
            "delete": {
                "tags": [
                    "Dictionary"
                ],
                "summary": "Delete dictionary by id",
                "operationId": "3d4883ecd167298247ae2a1ebb7f65cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "dictionary ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "e1288783-a5d4-1c4c-2cd6-0e92f7cc3bf9"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "dictionary permentally deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "dictionary permentally deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "protected route need to include sign in token as authorization bearer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "you need to include the authorization token from login"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "dictionary failed to permentally deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "dictionary not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/dct": {
            "post": {
                "tags": [
                    "Dictionary"
                ],
                "summary": "Post dictionary",
                "description": "Create a new dictionary using the given name and category. This request is using MySQL database.",
                "operationId": "4351a4a30e558c17f9b0c59424325998",
                "responses": {
                    "201": {
                        "description": "Dictionary created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/DictionaryResponse"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "dictionary created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "protected route need to include sign in token as authorization bearer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "you need to include the authorization token from login"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "failed"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "dictionary name must be at least 2 characters"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "failed"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "dictionary type must be one of the following values inventory_category, inventory_unit, inventory_room, reminder_type, reminder_context, report_category"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "failed"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "dictionary_name is a required field"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "failed"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "dictionary name has been used. try another"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/dct/{type}": {
            "get": {
                "tags": [
                    "Dictionary"
                ],
                "summary": "Show dictionary by type",
                "operationId": "0635b54e376376f26f000098c180b486",
                "responses": {
                    "200": {
                        "description": "dictionary found"
                    },
                    "401": {
                        "description": "protected route need to include sign in token as authorization bearer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "you need to include the authorization token from login"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "dictionary not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "dictionary not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "DictionaryResponse": {
                "properties": {
                    "dictionary_name": {
                        "type": "string",
                        "example": "Check"
                    },
                    "dictionary_type": {
                        "type": "string",
                        "example": "inventory_category"
                    },
                    "_id": {
                        "type": "string",
                        "example": "674342a8b53aabe966070f3d"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-24T15:13:44.621Z"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-24T15:13:44.621Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}